Простейшая отрисовка настроек аккаунта и тесты

+++ app/account/routes.py
@@ -17,18 +17,18 @@ class Account(FlaskView):
     @check_auth
     @route("/setting/", methods=["GET", "POST"])
     def set_account(self):
+        form = SettingsForm(request.form)
         uid, claims = get_auth_instance().get_current_user_data_from_token()
-        if request.method == "POST":     
-            #TODO: написать загрузку аватарки
-            form = SettingsForm()
-
-            if not form.validate_on_submit():
+        if request.method == 'POST':     
+            #TODO: написать загрузку аватарки            
+            
+            if not form.validate():
                 #в setting не забыть выводить сообщения об ошибке
-                return render_template("account/setting.html", form=form)
+                return render_template('account/setting.html', form=form)
 
             if form.delete.data:
-                return redirect("/delete"), 303
-
+                return redirect('/delete/'), 303
+                       
             user = User.query.filter_by(id = uid).first()
             user.nickname = form.nickname.data
             user.first_name = form.first_name.data
@@ -37,27 +37,36 @@ class Account(FlaskView):
             user.telegram_nickname = form.telegram_nickname.data
             db.session.commit()
             
-            return redirect("/setting"), 303
+            return redirect('/setting/'), 303
 
-        return render_template("account/setting.html", user = claims)   
+        form.nickname = claims.nickname
+        form.first_name = claims.first_name
+        form.last_name = claims.last_name
+        form.email = claims.email
+        form.telegram_nickname = claims.telegram_nickname
+
+        return render_template('account/setting.html', form = form)   
 
     @check_auth
     @route('/delete/', methods=["GET", "POST"])
     def delete_account(self):
-        form = DeleteFeedback()
+        form = DeleteFeedback(request.form)
 
-        if not form.validate_on_submit():
-            return render_template("account/delete.html", form=form)
+        if request.method == 'POST':
 
-        #куда сохраняем сообщение?
-        message = form.message.data
-        uid, claim = get_auth_instance().get_current_user_data_from_token()
+            if not form.validate_on_submit():
+                return render_template("account/delete.html", form=form)
 
-        user = User.query.filter_by(id = uid).first()
-        db.session.delete(user)
-        db.session.commit()
+            #куда сохраняем сообщение?
+            message = form.message.data
+            uid, claim = get_auth_instance().get_current_user_data_from_token()
 
-        return redirect("/home/"), 303
+            user = User.query.filter_by(id = uid).first()
+            db.session.delete(user)
+            db.session.commit()
+
+            return redirect("/home/"), 303
+        return render_template("account/delete.html", form=form)
 
 
 Account.register(bp)
